Split out the login modal
