+++ components/layout/Sidebar.tsx
@@ -1,26 +1,26 @@
-import styled from 'styled-components';
-import Logo from './Logo';
-import Nav from './Nav';
-import Title from './Title';
-import { Props } from '../../styles/Themes';
-import { motion } from 'framer-motion';
+import styled from "styled-components";
+import Logo from "./elements/Logo";
+import Nav from "./Nav";
+import Title from "./elements/Title";
+import { Props } from "../../styles/Themes";
+import { motion } from "framer-motion";
 
 const sidebarVariants = {
   enter: {
-    x: '0%',
+    x: "0%",
     opacity: 1,
     transition: {
       duration: 0.25,
-    }   
+    },
   },
   exit: {
-    x: '-100%',
+    x: "-100%",
     opacity: 0,
     transition: {
-      duration: 0.5
-    }
-  }
-}
+      duration: 0.5,
+    },
+  },
+};
 
 const StyledSidebar = styled(motion.aside)`
   position: relative;
@@ -33,7 +33,7 @@ const StyledSidebar = styled(motion.aside)`
     left: 0;
     width: 100%;
     height: 100%;
-    content: '\\A';
+    content: "\\A";
     z-index: -1;
   }
 
@@ -56,12 +56,12 @@ const StyledSidebarInner = styled.div`
 `;
 
 const Sidebar = () => (
-      <StyledSidebar variants={sidebarVariants}>
-        <StyledSidebarInner>
-          <Logo />
-          <Title />
-          <Nav />
-        </StyledSidebarInner>
-      </StyledSidebar>
-  );
+  <StyledSidebar variants={sidebarVariants}>
+    <StyledSidebarInner>
+      <Logo />
+      <Title />
+      <Nav />
+    </StyledSidebarInner>
+  </StyledSidebar>
+);
 export default Sidebar;
